lambdas/thumbnail/pyproject.toml
[project]
name = "t4_lambda_thumbnail"
version = "0.0.1"
requires-python = "~=3.13.0"

dependencies = [
  "numpy~=2.3",
  "pillow~=12.2",
  "python-pptx~=1.0.0",
  "requests>=2.26.0,<3",
  "t4_lambda_shared @ https://github.com/quiltdata/quilt/archive/d496dffbfb4b7a2ae05f6c1f7f0cb7d5d43bc984.zip#subdirectory=lambdas/shared",
  "bioio~=3.0",
  "bioio-czi~=2.4",
  "bioio-tifffile~=1.3",
  "bioio-ome-tiff~=1.4",
  "bioio-imageio~=1.3",
  "pypdfium2>=4,<5",
]

[build-system]
requires = ["uv_build~=0.9"]
build-backend = "uv_build"

[dependency-groups]
test = [
    "pytest~=9.0",
    "pytest-cov~=7.0",
    "pytest-mock~=3.15",
    "responses~=0.25",
    "quilt3~=7.0",
]
prod = [
    "awslambdaric~=3.1",
]

[tool.uv]
default-groups = ["test"]
